+++ src/main/java/net/sourceforge/jeuclid/element/MathPhantom.java
@@ -46,7 +46,7 @@ public class MathPhantom extends AbstractMathElementWithChildren {
 
     /** {@inheritDoc} */
     @Override
-    public void paint(final Graphics2D g, final int posX, final int posY) {
+    public void paint(final Graphics2D g, final float posX, final float posY) {
         super.paint(g, posX, posY);
     }
 
@@ -57,19 +57,19 @@ public class MathPhantom extends AbstractMathElementWithChildren {
 
     /** {@inheritDoc} */
     @Override
-    public int calculateAscentHeight(final Graphics2D g) {
+    public float calculateAscentHeight(final Graphics2D g) {
         return super.calculateChildrenAscentHeight(g);
     }
 
     /** {@inheritDoc} */
     @Override
-    public int calculateDescentHeight(final Graphics2D g) {
+    public float calculateDescentHeight(final Graphics2D g) {
         return super.calculateChildrenDescentHeight(g);
     }
 
     /** {@inheritDoc} */
     @Override
-    public int calculateWidth(final Graphics2D g) {
+    public float calculateWidth(final Graphics2D g) {
         return super.calculateChildrenWidth(g);
     }
